+++ core/src/main/java/org/geysermc/geyser/entity/type/TippedArrowEntity.java
@@ -47,13 +47,13 @@ public class TippedArrowEntity extends AbstractArrowEntity {
         int potionColor = entityMetadata.getPrimitiveValue();
         // -1 means no color
         if (potionColor == -1) {
-            dirtyMetadata.put(EntityDataTypes.CUSTOM_DISPLAY, 0);
+            dirtyMetadata.put(EntityDataTypes.CUSTOM_DISPLAY, (byte) 0);
         } else {
             TippedArrowPotion potion = TippedArrowPotion.getByJavaColor(potionColor);
             if (potion != null && potion.getJavaColor() != -1) {
                 dirtyMetadata.put(EntityDataTypes.CUSTOM_DISPLAY, (byte) potion.getBedrockId());
             } else {
-                dirtyMetadata.put(EntityDataTypes.CUSTOM_DISPLAY, 0);
+                dirtyMetadata.put(EntityDataTypes.CUSTOM_DISPLAY, (byte) 0);
             }
         }
     }
